requirements.txt
anyio==4.5.2
backports.zoneinfo==0.2.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dateparser==1.2.1
exceptiongroup==1.3.0
firebase-admin==6.9.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
msgpack==1.1.0
numpy==1.24.4
pandas==2.0.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
